Add files via upload
The goal of a fraud classification model is to accurately identify fraudulent activities or transactions from a large volume of data. Fraudulent activities can cause significant financial losses and damage to businesses, so it is crucial to develop effective models that can detect and prevent fraudulent behavior.

To build a fraud classification model, labeled data is required, consisting of both fraudulent and non-fraudulent examples. The model learns from this data to identify patterns and features that distinguish fraudulent transactions from legitimate ones.